engine-recommender-api/requirements.txt
transformers==4.38.2
sentence-transformers==2.6.1
keybert==0.8.3
huggingface_hub==0.20.3
scikit-learn==1.3.2
nltk==3.8.1
flask==2.3.3
python-dotenv==1.0.1
requests==2.31.0
pandas==2.1.4